backend/requirements.docker.txt
# Essential packages for RAI Chat backend
Flask==2.2.3
Werkzeug==2.2.3
Flask-Cors==3.0.10
SQLAlchemy==2.0.20
waitress==2.1.2
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
tavily-python==0.2.6
marshmallow==3.20.1
loguru==0.7.0
gunicorn==21.2.0
bcrypt
pydantic==1.10.8
email-validator
PyMySQL==1.1.0
cryptography==41.0.5
